+++ app/helpers/home_helper.rb
@@ -1,0 +1,2 @@
+module HomeHelper
+end
