+++ app/javascript/stylesheets/header.scss
@@ -41,7 +41,16 @@
 .search-logo {
   padding: 0 8px;
   position: absolute;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   right: 0;
+  cursor: pointer;
+  border-radius: var(--radius);
+  &:hover {
+    background-color: rgba(59, 73, 223, 0.2);
+  }
 }
 
 .logo {
@@ -160,4 +169,11 @@
 .header__email {
   font-size: 0.875rem;
   opacity: 0.75;
+}
+
+.createPost__title {
+  margin-left: 10px;
+  display: flex;
+  text-align: center;
+  align-items: center;
 }
