+++ app/views/header/_header.slim
@@ -1,11 +1,12 @@
 .header-container 
     .logo-search-bar
-        = link_to posts_path
+        = link_to home_index_path
             =image_tag("https://dev-to-uploads.s3.amazonaws.com/uploads/logos/resized_logo_UQww2soKuUsjaOGNB38o.png", alt: "logo",class: "logo")
-        .search-bar 
-            input type="text" placeholder="Search..."
+        =form_with url: search_path, method: :get do |f|
+            .search-bar 
+                =f.text_field :search, placeholder: "Search..."
                 .search-logo
-                    =image_tag("/assets/search-icon.svg", alt: "search icon")
+                    =f.submit "Search", type: :image, src: "/assets/search-icon.svg"
     - if user_signed_in?
         .rh__group_button
             = link_to 'Create post', new_post_path, class: "sign-up-button"
