+++ app/views/home/index.slim
@@ -1,0 +1,5 @@
+.contain
+  = render "/posts/posts", posts: @posts
+
+==pagy_bootstrap_nav(@pagy)
+
