+++ app/views/posts/index.slim
@@ -1,6 +1,8 @@
 .contain
-  = render "/posts/posts", posts: @posts
+  - if current_user.role == "admin" 
+    = render "/posts/status_posts", posts: @pending_posts
+  - else
+    = render "/posts/status_posts", posts: @posts 
 
-==pagy_bootstrap_nav(@pagy)
