+++ config/routes.rb
@@ -1,5 +1,6 @@
 Rails.application.routes.draw do
-  root to: 'users#root_routing'
+  root to: 'home#root_routing'
+  get 'home/index', to: "home#index"
   devise_for :users, path_names: {
     registration: 'register'
 }
@@ -12,6 +13,7 @@ Rails.application.routes.draw do
     # end
   end
   resources :users
-  get "archived", to: "posts#archive", as: "dashboard"
+  get "dashboard", to: "posts#index", as: "dashboard"
+  get "search", to: "home#search", as: "search"
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
